landing inputs looking clean
